big changes!! API/Extenders is moved to PlugIns/Extenders Server IIS settings was changed to the global whist site at my own domain RoundStatus CurrentPlayer and LeadingPlayer was changed to the Enum created yesterday

+++ Server/PlugIn/DumbPlayer.cs
@@ -63,7 +63,7 @@ namespace PlugIn
             Card? card = null;
             if (status.LeadingPlayer > 0)
             {
-                Suit s = status.CurrentPlay[status.LeadingPlayer].Value.Suit;
+                Suit s = status.CurrentPlay[(int)status.LeadingPlayer].Value.Suit;
                 card = (from cr in cards
                           where cr.Suit == s
                           select cr).FirstOrDefault();
